% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmt_dollar.R
\name{fmt_dollar}
\alias{fmt_dollar}
\title{Format Currency Values}
\usage{
fmt_dollar(x, digits = 2)
}
\arguments{
\item{x}{Numeric value to format}

\item{digits}{Number of decimal places (default: 2)}
}
\value{
Character string formatted as currency (e.g. "$1,234.56")
}
\description{
Format numeric values as US currency with proper symbols and comma separators.
Used throughout the app to display monetary amounts in a consistent,
human-readable format.
}
\examples{
fmt_dollar(1234.567)       # "$1,234.57"
fmt_dollar(1000000, digits = 0)  # "$1,000,000"
fmt_dollar(c(10, 20.5, 100))

}
\seealso{
Other utils:
\code{\link[=apply_rules]{apply_rules()}},
\code{\link[=fmt_pct]{fmt_pct()}}
}
\concept{utils}
