% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fmt_pct.R
\name{fmt_pct}
\alias{fmt_pct}
\title{Format Percentage Values}
\usage{
fmt_pct(x, digits = 1)
}
\arguments{
\item{x}{Numeric value to format (typically 0-100)}

\item{digits}{Number of decimal places (default: 1)}
}
\value{
Character string formatted as percentage (e.g. "12.3\%")
}
\description{
Format numeric values as percentages. Used throughout the app to display
ratios and percentages in a consistent, human-readable format.
}
\examples{
fmt_pct(30)           # "30.0\%"
fmt_pct(33.333, digits = 2)  # "33.33\%"
fmt_pct(c(10, 25, 50))

}
\seealso{
Other utils:
\code{\link[=apply_rules]{apply_rules()}},
\code{\link[=fmt_dollar]{fmt_dollar()}}
}
\concept{utils}
