% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_app.R
\name{launch_app}
\alias{launch_app}
\title{Launch the pffoundations application}
\usage{
launch_app(...)
}
\arguments{
\item{...}{Arguments passed to \code{\link[shiny:shinyApp]{shiny::shinyApp()}}}
}
\value{
A \code{shiny.appobj}
}
\description{
Launch the pffoundations application
}
\examples{
if (interactive()) {
  launch_app()
}
}
