% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_budgeting.R
\name{mod_budgeting_server}
\alias{mod_budgeting_server}
\title{Budgeting module server}
\usage{
mod_budgeting_server(id)
}
\arguments{
\item{id}{Module namespace ID}
}
\description{
Budgeting module server
}
