% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_budgeting.R
\name{mod_budgeting_ui}
\alias{mod_budgeting_ui}
\title{Budgeting module UI}
\usage{
mod_budgeting_ui(id)
}
\arguments{
\item{id}{Module namespace ID}
}
\value{
A \code{bslib::layout_sidebar} tag object
}
\description{
Budgeting module UI
}
