% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_emergency_fund.R
\name{mod_emergency_fund_server}
\alias{mod_emergency_fund_server}
\title{Emergency Fund module server}
\usage{
mod_emergency_fund_server(id)
}
\arguments{
\item{id}{Module namespace ID}
}
\description{
Emergency Fund module server
}
