% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_emergency_fund.R
\name{mod_emergency_fund_ui}
\alias{mod_emergency_fund_ui}
\title{Emergency Fund module UI}
\usage{
mod_emergency_fund_ui(id)
}
\arguments{
\item{id}{Module namespace ID}
}
\value{
A \code{bslib::layout_sidebar} tag object
}
\description{
Emergency Fund module UI
}
