% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_income.R
\name{mod_income_ui}
\alias{mod_income_ui}
\title{Income module UI}
\usage{
mod_income_ui(id)
}
\arguments{
\item{id}{Module namespace ID}
}
\value{
A \code{bslib::layout_sidebar} tag object
}
\description{
Income module UI
}
