% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_managing_debt.R
\name{mod_managing_debt_server}
\alias{mod_managing_debt_server}
\title{Managing Debt module server}
\usage{
mod_managing_debt_server(id)
}
\arguments{
\item{id}{Module namespace ID}
}
\description{
Managing Debt module server
}
