% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_managing_debt.R
\name{mod_managing_debt_ui}
\alias{mod_managing_debt_ui}
\title{Managing Debt module UI}
\usage{
mod_managing_debt_ui(id)
}
\arguments{
\item{id}{Module namespace ID}
}
\value{
A \code{bslib::layout_sidebar} tag object
}
\description{
Managing Debt module UI
}
