% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monthly_contribution_needed.R
\name{monthly_contribution_needed}
\alias{monthly_contribution_needed}
\title{Monthly contribution needed to reach an emergency fund goal}
\usage{
monthly_contribution_needed(target, current_savings, months)
}
\arguments{
\item{target}{Target emergency fund balance}

\item{current_savings}{Current savings already set aside}

\item{months}{Number of months in which to reach the target}
}
\value{
Required monthly contribution in dollars
}
\description{
Monthly contribution needed to reach an emergency fund goal
}
\examples{
monthly_contribution_needed(18000, 3000, 30)  # $500/mo
}
\seealso{
Other emergency_fund:
\code{\link[=ef_growth]{ef_growth()}},
\code{\link[=emergency_fund_target]{emergency_fund_target()}},
\code{\link[=months_to_ef_goal]{months_to_ef_goal()}}
}
\concept{emergency_fund}
