% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/months_to_break_even.R
\name{months_to_break_even}
\alias{months_to_break_even}
\title{Months to Break Even}
\usage{
months_to_break_even(upfront_cost, monthly_savings)
}
\arguments{
\item{upfront_cost}{Numeric upfront investment in dollars}

\item{monthly_savings}{Numeric recurring monthly savings in dollars}
}
\value{
Numeric number of months to break even (upfront_cost / monthly_savings)
}
\description{
Determine payback period for upfront costs with recurring monthly savings.
}
\examples{
months_to_break_even(500, 60)
months_to_break_even(2000, 150)

}
\seealso{
Other Budgeting:
\code{\link[=percent_of_income]{percent_of_income()}},
\code{\link[=to_annual]{to_annual()}},
\code{\link[=to_monthly]{to_monthly()}}
}
\concept{Budgeting}
