% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percent_of_income.R
\name{percent_of_income}
\alias{percent_of_income}
\title{Percent of Income}
\usage{
percent_of_income(expense, income)
}
\arguments{
\item{expense}{Numeric expense amount in dollars}

\item{income}{Numeric income amount in dollars}
}
\value{
Numeric percentage (0-100 or higher)
}
\description{
Calculate what share of income an expense represents.
}
\examples{
percent_of_income(1500, 5000)
percent_of_income(2000, 3000)

}
\seealso{
Other Budgeting:
\code{\link[=months_to_break_even]{months_to_break_even()}},
\code{\link[=to_annual]{to_annual()}},
\code{\link[=to_monthly]{to_monthly()}}
}
\concept{Budgeting}
