% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pffoundations-package.R
\docType{package}
\name{pffoundations-package}
\alias{pffoundations}
\alias{pffoundations-package}
\title{pffoundations: Personal Finance Foundations}
\description{
An interactive Shiny interface for the math and programming concepts covered in five personal finance chapters: income, tracking spending, budgeting, managing debt, and emergency funds.
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

Authors:
\itemize{
  \item First Last \email{first.last@example.com}
}

}
\keyword{internal}
