% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{std_categories}
\alias{std_categories}
\title{Standard transaction category names}
\format{
A character vector of length 14.
}
\usage{
std_categories
}
\description{
A character vector of the 14 category labels used throughout the app,
including \code{"Other"} for transactions that do not match any rule in
\link{default_rules}.
}
\seealso{
\link{default_rules}, \link{example_data}

Other tracking_spending:
\code{\link[=apply_rules]{apply_rules()}},
\code{\link{default_rules}},
\code{\link{example_data}}
}
\concept{tracking_spending}
\keyword{datasets}
